+++ src/net/fe/Player.java
@@ -17,7 +17,7 @@ public class Player implements Serializable {
 	private Party party;
 	
 	/** The client id. */
-	private byte clientID;
+	private int clientID;
 	
 	/** The nickname. */
 	private String nickname;
@@ -46,7 +46,7 @@ public class Player implements Serializable {
 	 * @param name the name
 	 * @param id the id
 	 */
-	public Player(String name, byte id) {
+	public Player(String name, int id) {
 		party = new Party();
 		clientID = id;
 		nickname = name;
@@ -77,7 +77,7 @@ public class Player implements Serializable {
 	 *
 	 * @return the id
 	 */
-	public byte getID() {
+	public int getID() {
 		return clientID;
 	}
 	
@@ -95,7 +95,7 @@ public class Player implements Serializable {
 	 *
 	 * @param id the new client id
 	 */
-	public void setClientID(byte id) {
+	public void setClientID(int id) {
 		clientID = id;
 	}
