+++ src/net/fe/overworldStage/OverworldChat.java
@@ -87,9 +87,8 @@ public class OverworldChat extends TextInputBox implements DoNotDestroy {
 	 */
 	public void send() {
 		if(input.length() == 0) return;
-		byte id = FEMultiplayer.getClient().getID();
-		FEMultiplayer.getClient().sendMessage(
-				new ChatMessage(id, input.toString()));
+		int id = FEMultiplayer.getClient().getID();
+		FEMultiplayer.getClient().sendMessage(new ChatMessage(id, input.toString()));
 		input.delete(0, input.length());
 		cursorPos = 0;
 	}
